requirements.txt

specialsauce
gunicorn
psycopg2-binary
redis

alembic
decorator
dnspython
email-validator
Flask
Flask-Login
Flask-Migrate
Flask-SQLAlchemy
Flask-WTF
Jinja2
MarkupSafe
oauthlib
python-dateutil
python-dotenv
requests
requests-oauthlib
SQLAlchemy
urllib3
Werkzeug
WTForms